# docs/swagger.yaml
openapi: 3.0.0
info:
  title: yanbing-edge
  description: 接口文档,可能有问题
  version: 1.0.0
paths:
...
  /device-group:
    post:
      summary: 创建设备组
      operationId: createDeviceGroup
      tags:
        - 设备组管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceGroup'
      responses:
        '200':
          description: 成功创建设备组
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RDeviceGroup'
...
components:
  schemas:
...
    CreateDeviceGroup:
      type: object
      properties:
        name:
          type: string
          description: 设备组名称
        interval:
          type: integer
          description: 间隔时间
        device_id:
          type: integer
          description: 设备 ID
    DeviceGroup:
      type: object
      properties:
        id:
          type: integer
          description: 设备组 ID
        name:
          type: string
          description: 设备组名称
        interval:
          type: integer
          description: 间隔时间
        device_id:
          type: integer
          description: 设备 ID
    RDeviceGroup:
      type: object
      properties:
        code:
          type: integer
          description: 响应代码
        success:
          type: boolean
          description: 请求是否成功
        msg:
          type: string
          description: 响应消息
        data:
          $ref: '#/components/schemas/DeviceGroup'
...